requirements.txt
RPi.GPIO
pygame